Update on scroll element inner size change
When it didn’t trigger a size change of the observed scrollable element
(depends on site layout), then overflow indicators did not update on
scroll width/height change caused by a child node change.

For example: Content is added or removed dynamically, and it only
changed the inner size.

This commit handles this case by also updating overflow indicators on
size change of the scrollable element’s direct child nodes, which most
likely means a change of the scroll width or height.